+++ demo.py
@@ -7,7 +7,7 @@ from rapidocr import RapidOCR, VisRes
 
 from rapid_table import RapidTable, RapidTableInput, VisTable
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     # Init
     ocr_engine = RapidOCR()
     vis_ocr = VisRes()
@@ -19,8 +19,7 @@ if __name__ == '__main__':
     img_path = "tests/test_files/table.jpg"
 
     # OCR
-
-    rapid_ocr_output = ocr_engine(img_path, return_word_box=True)
+    rapid_ocr_output = ocr_engine(img_path)
     ocr_result = list(
         zip(rapid_ocr_output.boxes, rapid_ocr_output.txts, rapid_ocr_output.scores)
     )
@@ -32,16 +31,29 @@ if __name__ == '__main__':
     # ]
     # table_results = table_engine(img_path, ocr_result)
 
-    table_html_str, table_cell_bboxes = table_results.pred_html, table_results.cell_bboxes
+    table_html_str, table_cell_bboxes = (
+        table_results.pred_html,
+        table_results.cell_bboxes,
+    )
     # Save
     save_dir = Path("outputs")
     save_dir.mkdir(parents=True, exist_ok=True)
 
     save_html_path = save_dir / f"{Path(img_path).stem}.html"
-    save_drawed_path = save_dir / f"{Path(img_path).stem}_table_vis{Path(img_path).suffix}"
-    save_logic_points_path = save_dir / f"{Path(img_path).stem}_table_col_row_vis{Path(img_path).suffix}"
+    save_drawed_path = (
+        save_dir / f"{Path(img_path).stem}_table_vis{Path(img_path).suffix}"
+    )
+    save_logic_points_path = (
+        save_dir / f"{Path(img_path).stem}_table_col_row_vis{Path(img_path).suffix}"
+    )
 
     # Visualize table rec result
-    vis_imged = viser(img_path, table_results, save_html_path, save_drawed_path, save_logic_points_path)
+    vis_imged = viser(
+        img_path,
+        table_results,
+        save_html_path,
+        save_drawed_path,
+        save_logic_points_path,
+    )
 
     print(f"The results has been saved {save_dir}")
