+++ rapid_table/utils/__init__.py
@@ -1,3 +1,8 @@
 # -*- encoding: utf-8 -*-
 # @Author: SWHL
 # @Contact: liekkaskono@163.com
+from .download_model import DownloadModel
+from .load_image import LoadImage
+from .logger import Logger
+from .utils import is_url
+from .vis import VisTable
